Refactoring: move freshness, license and usability viewer to a tsx file
